src/requirements.txt

fastapi~=0.104.0
PyJWT~=2.8.0
starlette~=0.27.0
loguru~=0.7.2
APScheduler~=3.10.4
SQLAlchemy~=2.0.22
pydantic~=2.4.2
pymysql~=1.1.0
tortoise-orm~=0.20.0
python-multipart~=0.0.6
aiofiles~=23.2.1
aioredis~=2.0.1
aiomysql~=0.2.0
passlib~=1.7.4
itsdangerous~=2.1.2
bcrypt~=4.0.1
six~=1.16.0
httpx~=0.25.0
uvicorn~=0.23.2
pycryptodome~=3.19.0
icalendar~=5.0.10
jicson~=1.0.2
caldav~=1.3.6
requests~=2.31.0
pytz~=2023.3.post1
furl~=2.1.3
ujson~=5.9.0
marshmallow~=3.20.2
loguru~=0.7.2
gmssl~=3.2.2
dill~=0.3.8
rpyc~=5.3.1
aiohttp~=3.9.3
python-dotenv
pydantic-settings~=2.1.0
shortuuid~=1.0.11
httpx~=0.26.0
cryptography~=42.0.5